+++ integration/harness.rs
@@ -6,6 +6,7 @@ test_mods! {
     mod canonicalize;
     #[cfg(target_os = "linux")]
     mod consistent_id_mappings;
+    mod delete_before_lockdown;
     mod env;
     mod exec;
     mod exec_symlinked_dir;